moved return types to orchestators folder
